+++ packages/mdx/dev/content/comment-annotations.mdx
@@ -129,7 +129,13 @@ body {
 ```
 
 ```mdx
-# Hello {/* mark */}
+{/* mark(2) */}
+
+# Lorem
+
+## Foo
+
+# Ipsum {/* mark */}
 ```
 
 ```jsonc
