+++ packages/mdx/dev/content/simple-code.mdx
@@ -14,6 +14,13 @@ graph LR
     --> End1(End)
 ```
 
+```foobar
+// unknown lang
+function lorem(ipsum, dolor = 1) {
+  return ipsum + dolor;
+}
+```
+
 ```js
 function lorem(ipsum, dolor = 1) {}
 ```
