+++ packages/mdx/src/index.tsx
@@ -1,3 +1,1 @@
 export { attacher as remarkCodeHike } from "./remark/transform"
-
-export { highlight } from "./highlighter"
